HBP-90. Move DatatypeMeasureH5 under  core/entities/file/simulator to avoid service -> adapter dependency.
