+++ framework_tvb/tvb/tests/framework/adapters/analyzers/adapters_test.py
@@ -29,6 +29,7 @@
 #
 
 import os
+
 from tvb.adapters.analyzers.cross_correlation_adapter import CrossCorrelateAdapter, PearsonCorrelationCoefficientAdapter
 from tvb.adapters.analyzers.fcd_adapter import FunctionalConnectivityDynamicsAdapter
 from tvb.adapters.analyzers.fmri_balloon_adapter import BalloonModelAdapter
@@ -41,12 +42,12 @@ from tvb.adapters.analyzers.pca_adapter import PCAAdapter
 from tvb.adapters.analyzers.wavelet_adapter import ContinuousWaveletTransformAdapter
 from tvb.adapters.datatypes.h5.fcd_h5 import FcdH5
 from tvb.adapters.datatypes.h5.graph_h5 import CovarianceH5, CorrelationCoefficientsH5
-from tvb.adapters.datatypes.h5.mapped_value_h5 import DatatypeMeasureH5
 from tvb.adapters.datatypes.h5.mode_decompositions_h5 import PrincipalComponentsH5, IndependentComponentsH5
 from tvb.adapters.datatypes.h5.spectral_h5 import WaveletCoefficientsH5, CoherenceSpectrumH5, \
     ComplexCoherenceSpectrumH5
 from tvb.adapters.datatypes.h5.temporal_correlations_h5 import CrossCorrelationH5
 from tvb.adapters.datatypes.h5.time_series_h5 import TimeSeriesRegionH5
+from tvb.core.entities.file.simulator.datatype_measure_h5 import DatatypeMeasureH5
 from tvb.core.neocom import h5
 from tvb.tests.framework.core.base_testcase import TransactionalTestCase
