+++ framework_tvb/tvb/tests/framework/core/services/burst_service_test.py
@@ -38,10 +38,10 @@ from time import sleep
 import numpy
 from tvb.adapters.datatypes.db.mapped_value import DatatypeMeasureIndex
 from tvb.adapters.datatypes.db.simulation_history import SimulationHistoryIndex
-from tvb.adapters.datatypes.h5.mapped_value_h5 import DatatypeMeasureH5
 from tvb.config.init.introspector_registry import IntrospectionRegistry
 from tvb.core.adapters.abcadapter import ABCAdapter
 from tvb.core.entities.file.files_helper import FilesHelper
+from tvb.core.entities.file.simulator.datatype_measure_h5 import DatatypeMeasureH5
 from tvb.core.entities.model.model_burst import *
 from tvb.core.entities.model.model_datatype import *
 from tvb.core.entities.model.model_operation import STATUS_STARTED, OperationPossibleStatus, STATUS_FINISHED
